Extend Storage classes instead of implementing interfaces
